+++ code/defaults.php
@@ -6,9 +6,9 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 class GDBTODefaults {
 	var $default_options = array(
-		'version'                                => '3.5.2',
-		'date'                                   => '2024.05.15.',
-		'build'                                  => 2440,
+		'version'                                => '3.5.3',
+		'date'                                   => '2024.08.19.',
+		'build'                                  => 2450,
 		'status'                                 => 'stable',
 		'product_id'                             => 'gd-bbpress-tools',
 		'edition'                                => 'free',
@@ -62,6 +62,6 @@ class GDBTODefaults {
 		'd4p_bbpt_signature_enhanced',
 	);
 
-	function __construct() {
+	public function __construct() {
 	}
 }