+++ forms/more/toolbox.php
@@ -10,15 +10,14 @@ if ( ! defined( 'ABSPATH' ) ) {
     <fieldset class="the-top">
         <em><?php _e( 'consider upgrading to', 'gd-bbpress-tools' ); ?></em>
         <h3>
-            <a title="GD bbPress Toolbox Pro" target="_blank" href="<?php echo d4p_url_campaign_tracking( 'https://plugins.dev4press.com/gd-bbpress-toolbox/', 'gd-bbpress-tools-to-pro', 'wordpress', 'free-plugin-admin' ); ?>">GD bbPress Toolbox Pro</a>
+            <a title="GD bbPress Toolbox Pro" target="_blank" href="<?php echo d4p_url_campaign_tracking( 'https://www.dev4press.com/plugins/gd-bbpress-toolbox/', 'gd-bbpress-tools-to-pro', 'wordpress', 'free-plugin-admin' ); ?>">GD bbPress Toolbox Pro</a>
         </h3>
         <p style="font-weight: bold; font-size: 15px;"><?php _e( 'Enhancing WordPress forums powered by bbPress', 'gd-bbpress-tools' ); ?></p>
         <p><?php _e( 'Expand bbPress powered forums with attachments upload, BBCodes, signatures, widgets, quotes, topics report, thanks, activity tracking, enhanced widgets, extra views...', 'gd-bbpress-tools' ); ?></p>
 
         <h4><?php _e( 'Important links', 'gd-bbpress-tools' ); ?></h4>
-        <a href="<?php echo d4p_url_campaign_tracking( 'https://plugins.dev4press.com/gd-bbpress-toolbox/', 'gd-bbpress-tools-to-pro', 'wordpress', 'free-plugin-admin' ); ?>" target="_blank"><?php _e( 'Official Plugin Website', 'gd-bbpress-tools' ); ?></a><br/>
-        <a href="<?php echo d4p_url_campaign_tracking( 'https://plugins.dev4press.com/gd-bbpress-toolbox/articles/toolbox-pro-vs-free-plugins/', 'gd-bbpress-tools-to-pro', 'wordpress', 'free-plugin-admin' ); ?>" target="_blank"><?php _e( 'Toolbox Pro vs Free plugins', 'gd-bbpress-tools' ); ?></a><br/>
-        <a href="<?php echo d4p_url_campaign_tracking( 'https://club.dev4press.com/', 'gd-bbpress-tools-to-pro', 'wordpress', 'free-plugin-admin' ); ?>" target="_blank"><?php _e( 'Dev4Press Plugins Club Membership', 'gd-bbpress-tools' ); ?></a><br/>
+        <a href="<?php echo d4p_url_campaign_tracking( 'https://www.dev4press.com/plugins/gd-bbpress-toolbox/', 'gd-bbpress-tools-to-pro', 'wordpress', 'free-plugin-admin' ); ?>" target="_blank"><?php _e( 'Official Plugin Website', 'gd-bbpress-tools' ); ?></a><br/>
+        <a href="<?php echo d4p_url_campaign_tracking( 'https://www.dev4press.com/dev4press-club/', 'gd-bbpress-tools-to-pro', 'wordpress', 'free-plugin-admin' ); ?>" target="_blank"><?php _e( 'Dev4Press Plugins Club Membership', 'gd-bbpress-tools' ); ?></a><br/>
     </fieldset>
 
     <fieldset>